Update xterm.js and addon-fit to latest versions
xterm.js dependency was getting quite old, version-wise.
There was a recent issue with the latest Safari losing focus during the
login, and so it was worth a try to see if a refresh fixed it.
xterm.js is distributed via npm, so embraced this and removed local copies.
Moved the combined libraries into ./lib, rather than ./js for neatness.
New xterm.js needed changes with removal of setOptions and handling
the plugin.