Safari hack to stop it silently ignoring passwords
Not sure the issue, but Safari stopped reading keydown events
after we return too many false values from the event handler,
hiding keypress and keydown, but not keyup.
Returning false mostly just made it look right, but the
onData event doesnt take keys anyway till the connection is fully made,
so if it gets a few extra keys before that it makes no difference.
Look to change it back in the future if the bug is fixed, or the issue
is better understood.