Basic fix for ssh-rsa deprecation
Host keys could previously be verified using ssh-rsa, but
this was deprecated a while ago due to security concerns. Those
concerns do not matter to the basic sshy operation, as host keys
are not required for username/password authentication. But we do
need to agree on a method with sshd so we could ask for them.
For now, simply advertise support for rsa-sha2-256.